Switch to Postgres (#39)
* WIP Switch to Postgres

* feat(postgres): more work on porting to postgres, now compiles

* feat(docker-compose): Changed the docker-compose.yml file to use postgres.

* Update docker, documentation, gh actions...

* Remove bson dependency

* Remove bson import

* feat: move mock filehost to trait rather than cargo feature

* feat(postgres): transactions for mod creation, multipart refactor

* fix: Add Cargo.lock so that sqlx functions

* Update sqlx offline build data

* fix: Use SQLX_OFFLINE to force sqlx into offline mode for CI

* Default release channels

* feat(postgres): refactor database models to fit postgres models

* fix: Fix sqlx prepare, fix double allocation in indexing

* Add dockerfile (#40)

+++ src/search/indexing/curseforge_import.rs
@@ -178,7 +178,7 @@ pub async fn index_curseforge(
             .replace("/256/256/", "/64/64/");
 
         docs_to_add.push(SearchMod {
-            mod_id: -curseforge_mod.id,
+            mod_id: -curseforge_mod.id as i64,
             author: (&curseforge_mod.authors[0].name).to_string(),
             title: curseforge_mod.name,
             description: curseforge_mod.summary.chars().take(150).collect(),
