+++ src/search/mod.rs
@@ -47,7 +47,7 @@ impl actix_web::ResponseError for SearchError {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct SearchMod {
-    pub mod_id: i32,
+    pub mod_id: i64,
     pub author: String,
     pub title: String,
     pub description: String,
@@ -66,7 +66,7 @@ pub struct SearchMod {
 }
 
 impl Document for SearchMod {
-    type UIDType = i32;
+    type UIDType = i64;
 
     fn get_uid(&self) -> &Self::UIDType {
         &self.mod_id
